Commented slurm and python script for tsne
I added the 2 scripts I've been using to run tsne in the hpcc. I gave indications of type of input files. Also I added a command in hpcc_script2 to save tsne object as a coordinate file and in the slurm_script to save python output (verbose) into a file.